introduced ModifierBase class created new Modifier namespace update all tests
