+++ Base/IFsuipcConfigItem.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using MobiFlight;
+using MobiFlight.Modifier;
 using MobiFlight.OutputConfig;
 
 namespace MobiFlight
