+++ FSUIPC/FSUIPCConfigItem.cs
@@ -1,4 +1,5 @@
-﻿using MobiFlight.OutputConfig;
+﻿using MobiFlight.Modifier;
+using MobiFlight.OutputConfig;
 using System;
 using System.Collections.Generic;
 using System.Linq;
