+++ UI/Panels/Config/InterpolationPanel.cs
@@ -6,6 +6,7 @@ using System.Data;
 using System.Linq;
 using System.Text;
 using System.Windows.Forms;
+using MobiFlight.Modifier;
 
 namespace MobiFlight.UI.Panels.Config
 {
